create gain.rs
